// frontend/src/api/model/concourseItemVersionRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Libre-Q API
 * OpenAPI spec version: 0.1.0
 */
import type { ConcourseItemStatus } from './concourseItemStatus';
import type { ConcourseItemVersionReadSource } from './concourseItemVersionReadSource';
import type { ConcourseItemVersionReadTranslationsSnapshotItem } from './concourseItemVersionReadTranslationsSnapshotItem';
import type { ConcourseItemVersionReadChangeComment } from './concourseItemVersionReadChangeComment';
import type { ConcourseItemVersionReadChangedBy } from './concourseItemVersionReadChangedBy';

/**
 * Schema for reading a concourse item version snapshot.
 */
export interface ConcourseItemVersionRead {
    id: number;
    item_id: number;
    version_number: number;
    code: string;
    status: ConcourseItemStatus;
    source?: ConcourseItemVersionReadSource;
    translations_snapshot?: ConcourseItemVersionReadTranslationsSnapshotItem[];
    tag_ids_snapshot?: number[];
    change_comment?: ConcourseItemVersionReadChangeComment;
    changed_by?: ConcourseItemVersionReadChangedBy;
    changed_at: string;
}